--- angular-universal-ssr/client-app/src/app/client-weather/client-weather-city/client-weather-city.component.ts
@@ -1,15 +1,0 @@
-import { Component, Input, OnInit } from "@angular/core";
-
-@Component({
-  selector: "app-client-weather-city",
-  templateUrl: "./client-weather-city.component.html",
-  styleUrls: ["./client-weather-city.component.scss"],
-})
-export class ClientWeatherCityComponent implements OnInit {
-  @Input()
-  public city: string;
-
-  constructor() {}
-
-  ngOnInit(): void {}
-}
